[GATEWAY] Sequence handling

+++ discord_client_gateway/src/events/structs/gateway.rs
@@ -6,7 +6,7 @@ use serde_json::Value;
 pub struct GatewayPayload {
     pub op: u8,
     pub d: Value,
-    pub s: Option<u64>,
+    pub s: Option<u32>,
     pub t: Option<String>,
 }
